backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

pandas==2.1.3
openpyxl==3.1.2
twilio==8.10.0
sendgrid==6.12.5
